Write a class for handling incoming data and class for storing data
DataStore wraps numpy arrays to store the incoming data in oversized
arrays, preventing unneccessary data copying to resize numpy arrays.

This should be compared to using numpy arrays directly, as I don't know
if this is necessary. std::vector would use the behaviour of DataStore
directly, always overallocating to a factor of 2. Numpy may or may not
do that.

DataMediator owns and connects the datastore and datasource (socketclient) to
each other.