Create a basic ui for connecting/disconnecting
Consists of a toolbar of 4 lineedits to input ipv4 ip addresses, a
spinbox for the port number a connect/disconnect button. The lineedits
currently allow for 3-digit inputs and a + symbol. They should be
limited to 0-255 with no symbols, but the base validator seems
insufficient (at least without further customization of it)

A wrapper class tracks the ip value as a string. It does not do any
additional validation and will use the contents of the ip fields
directly.