Remove old datastore and use pandas
Remove the old datastore and download logic in favor of the validated
dataframe. For the moment we still send numpy arrays to matplotlib. This
stops us from using the unit for anything useful, but kept the changes
smaller.

Time Of Day widget creates a dataframe from the base arrays to calculate
the buckets, so we might want to send the relevant data to the plots in
the future to simplify things like that.