Update schema to be in line with the new server version
The server was updated to use a schema with categorical dtypes. Issues
with the index during serialization with the categorical dtypes means we
can't rely on the multiindex being correct. Instead the index is
intentionally dropped and rebuilt after being loaded. The plotter then
needs to know how to rebuild it.

The SensorReading had the indexes in the wrong order, fixed.