Add notebooks for line and stat-plots
Add some notebooks for drawing the line plots of single datastreams and
for drawing time-of-day grouped data for the single datastreams. The
time-of-day grouped data is drawn as maximums, minimums and mean.

Includes implementation in both pandas and polars.
Polars is faster at the stats one, while pandas wins for
lineplots. Using the builtin plot method for pandas for the lineplot is
exceptionally slow, as it does timestamp conversions back and forth for
some reason. Plotting numpy arrays directly is much faster. Polars
spends considerable time in selection that pandas handles faster.
Perhaps the selection logic can be written to be faster.

Both libraries are faster at the time-of-day drawing than simply drawing
the raw data.