Move plotting responsibility to abstract plotmanager
Moves the plotting responsibility out of the QT class that owns the
figure and canvas that houses the plot. The QT class instead exposes an
interface for the manager to create and remove subplots in the figure.

Different managers can be created that organize plots differently.
Simple ones that only allow 1 Axes and 1 Artist/PlotStrategy, or 1 Axes
and several artists in that axes, or multiple axes that organise in
specific ways.

The manager is aware of the underlying widget, so it can work with the
Figure and query the widget about things like isVisible.