Manager that makes a new axes for each strategy
Implement a simple mangager that just draws each strategy into a
separate axes. There is no layout management for the separate axes, so
they just draw on top of each other, with the latest one on top. Without
layout-management it's only really useful for a widget that plots a
single thing, like the old UI-structure was doing.