A more complex manager that groups sensor-type plots together
This takes two keys when adding a plot-strategy. One for the axes it
goes to, and one for the strategy itself. This supports the SensorType,
Sensor as keys, and groups the same kind of measurements into a single
plot, while allowing you to add another plot for another kind of
measurement.

It has a basic layout management that definetly doesn't work as the user
might expect, and is mostly useful to experiment with for development.