Remove code duplication and align interfaces
The different managers have different number of required arguments when
adding/removing plots. This allows all of them to accept unlimited
arguments. This opens up for easier to misuse the interface, by making
it more forgiving, but decouples it a bit more from the caller.

Remocves an area of code duplication. When removing a plot, the artist
should be removed from the axes. The way this is done is the same for
all managers, so it's added into the abstract baseclass as a helper
function.

Adds some docstrings to the managers