Make package from raspberry sensor datahandling
Initially move the datatypes and remotereader into here.

The dataframehandler will also be added, but needs a good chunk of work,
as it now needs to be abstracted behind an interface.

The datatypes and remotereader modules are implementation details for
the raspberry sensors and can be quickly just added without further
work.