Interface for accessing a plottable dataset.
This defines an abstract FrameHandler. The FrameHandler manages a
dataframe and provides read-access to the dataframe, methods for loading
and updating the data in the dataframe, and a model that maps
plotstrategies to datasets that support that plotstrategy.

For instance, the dataset that conforms to index ("temperature",
"DHT11", :) supports being plotted as a line.

This way of providing direct read access and keys is a bit
implementation-leaky, and might be changed as new plotstrategies are
added. It can probably be vastly improved.