Reuse code for models
Humidity and Temperature models are identical apart from name. They're
both just one-dimensional time series, and can be plotted the same way,
with the same signature for data-fn.

This is where the nominal datatype abstractions come in. The models are
identical in structure and behaviour, but they are declared different in
name, so they are different.