Builder-pattern for navigation toolbar
The toolbar below the plots containing the Matplotlib navigation widget
contains additional custom widgets that makes sense for different
plot-widgets. Some of these widgets were added in the base-class
(drawwidget), some in the derived class.

This is quite inflexible. Removing a widget we don't want when added in
the baseclass is difficult and annoying. Manually creating the entirety
of the toolbar in every baseclass is copy-pasting code.

This moves creating the toolbar widget into a builder-class. The widgets
can now specify with a chained call to the builder how the
toolbar-widget should look.

The builder supports adding widgets to either the left or right edge of
the widget. The order they appear depends on the order of the
builder-call.