Factory pattern for plotwidget
The subclasses for the plot/draw widgets are all very similar. They
differ in the navigation toolbar, and in which plotstrategy they
implement. The code for them mentions the strategy-type they use. At the
same time, this was not visible at the point the widgets get created,
and an extra dictionary lookup was used to determine which widget was to
be made for which strategy. The dictionary was awkwardly placed.

This removes the entire subclass hierarchy. The hierarchy now is
DrawWidget, which creates the figure, the navigation toolbar, and
methods for creating and removing axes, and the PlotWidget which holds
the PlotManager and the DataTypeModel and PlotStrategy. It has methods
for calling the manager, and the manager will call methods in the
DrawWidget.

The individual widgets are now built by a factory instead. The previous
dictionary lookup is turned into a match-block that calls specific
factory methods. The factory methods create a plotwidget with specific
configurations. It creates the appropriate manager-type, although only 1
of these exists at the moment. It sends in the type of plotstrategy the
manager should be creating. It configures the navigation toolbar
builder. It then returns a PlotWidget that is configured appropriately.
The widgets now have different state, but the same type.