Use the new patterns when creating plottabs
This finally calls the new pattern for creating the plotwidgets at the
point of PlotTabWidget.

A lookup into a function that returns the plottypes a certain model
supports is made for each model. It then creates a PlotTabWidget which
has a sidebar for choosing which data to plot, and the PlotWidget, which
it now builds by calling to the PlotWidgetFactory. It then glues the
sidebar and plotwidget together and puts them into the layout.