System for drawing plot legends
This adds a strategy based system for drawing legends in plots. 3
concrete implementations are implemented. It can either draw a legend
internal to the Axes, or a legend for the whole figure. The latter gets
placed outside the Axes, where there is more space, but is less suited
in a plot with multiple sub-axes.

The third option is to create a colorbar that shows the colormap used.
This is useful for the timeofday plot. Matplotlib implements colormaps
differently to legends. It's basically a plot with its own subaxes that
gets added to the figure, while the legend is more like an Artist that
either floats in the figure, or is attached to an Axes.