Container class for plots, maps to axes and labels
Place plotstrategies in a container that enables and disables them as
requested. Keeps track of axes the plot uses. PlotStrategies
theoretically exist independent of axes and can just be given a
different axes if desired, so this behaviour might be split into
different strategies too. More cases for it is needed.