Remove tick management from the plotstrategies.
The strategies used to do some ticker-management on their own. But this
meant the line-plot was specialized to use datetime on the x-axes.

The time-of-day plot handled all its polar projection magic in the
ticker-management as well, hardcoding it to a polar projection. Most of
the code works the same for a flat projection, it just needs different
ticks. Decoupling the plot-strategy and ticker management lets us use
the time-of-day strategy to plot a completly different looking flat
plot.