Move yr locations and api secrets to toml files
The client-id is in a git-ignored file in the same directory as main
called frost_secrets.toml

The file sources.toml contains information to map locations to yr/frost
api calls to gather data for those locations. For frost/historic this is
simply a mapping to a sensor id, and for the yr forecast it's
coordinates and altitude, as the settings examplify. This way new
locations can be added, removed or tested without changing any code.