Remove references to PolarAxes in TimeOfDay plot
This plot can easily reuse this code to plot a flat time-of-day with
0-24 on the x-axis without changing anything in the time-of-day plot.

There is still an artifact of the polar plot in the time-of-day code
where the x-axis goes from 0 to 2 pi. This can be then intercepted by
the tick-strategy to be changed to a 0-24 range. This currently doesn't
work well if you start changing the graph interactively though.

It also adds back in the line actually defining __call__, which went
missing in the checked in code base, but was always there in my working
stash.