Make DataTypeModel a QObject for signals
This conflicts with ABC, so abstractmethods have been swapped with
NotImplementedError.

This change allows us to leverage the full signals and slots
event-handling from QT to forward information about when the dataloader
threads have finished getting data straight to the widgets that plot
them, and determine if they need to refresh.

Basically, even if there is new data and the widget is visible, we don't
need to do anything if that data isn't already plotted. The widget can
use these signals to see which datasets are changed, check if they're
plotted and only redraw as needed.