Fix making duplicate sources in sidebar
The sidebar got multiple signals that a source was added, one for each
dataline in the source. It therefore added that source multiple times.

Now it gets a fully qualified signal with both source and dataline
identifier. If the source doesn't exist, it makes it and adds the
dataline child to it. When the source does exist it fetches it and adds
the new child to it.